tests/wpt/web-platform-tests/app-history/navigate/return-value/goTo-current.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<script>
promise_test(async t => {
  assert_equals(appHistory.entries().length, 1);
  const entry = appHistory.current;

  const result = appHistory.goTo(appHistory.current.key);
  await assertBothFulfill(t, result, entry);

  assert_equals(appHistory.entries().length, 1);
  assert_equals(appHistory.current, entry);
}, "goTo() with current key");
</script>
